Add layer-wise chunking approach (Deprecated)
While the chunks are correctly generated using this appraoch, this
approach was oblivious to whether a tensor dimension was a row or
a column, and therefore this was incompatible with the zeroing func-
tionality which requires the exact location of any parameter we
want to zero out. The functions are added just for reference.