Add callback fn for ddmin_test in run_defect.py (UNDER TEST)
This is a function that is to be called back repeatedly by the
ddmin test method. The same function can be added in the other
model execution scripts (run_clone.py, run_gen.py, etc.). A refe-
rence to this fn would need to be sent to the module that runs
the ddmin (in our case, model_anacomp). See next commit.